#40 Fetch legal URLs from registry to show in bottom page fragment
The JSP fragment `bottom.jsp` is included in almost all pages that CAS
delivers. Previously, this bottom fragment contained two legal links to
imprint and the privacy policy.

This commit changes this behavior so 3 links from the registry can be
shown. Additional to the existing two links the new link sequence is now:

1. Terms of Service
2. Imprint
3. Privacy Policy

If one or more entries from the registry is missing then no links
will be shown. Entries not to be found are considered empty. Empty
entries are not rendered. Likewise, potentially superfluous link
delimiters may be subject of being not rendered.

Since registry access from the CAS is a rather expensive operation and
while the legal URLs are not supposed to change frequently these 3 links
are cached within the LegalLinkProducer bean.