#49 Fix shell script error handling, add startup logs; set error state
This commit fixes previous shell script errors, namely
- handle an unset root log level registry key
- fix misnamed variable
- handle `source` error more appropriately

That last one was tricky: The sourced `logging.sh` fails immediately if any
error occurs inside the logging script (working as intended). Anyhow, the
result of the sourcing command then leads to an non-zero exit code which in
turn sends the container into a restart loop.

This commit works around the `source` exit code by catching, and handling it.
The handling will set a dogu state along with orderly exiting the dogu startup
so administrators can understand the source of problem more quickly.

Setting the variable `loggingExitCode` prior to executing the logging script
serves to avoid an "unbound variable error". Otherwise the variable would be
created conditionally leading to an unbound state.