Add configuration of legal URLs #91
The URLs are configured in the etcd. You can configure one URL each for the terms of use, the imprint and the privacy policy.
Add configuration of legal URLs #91

The URLs are configured in the etcd. You can configure one URL each for the terms of use, the imprint and the privacy policy.