Change comment #99
Link added to origin code.

The code of 2 methods has been copied from the class
LdapAuthenticationConfiguration. This class comes from the cas library.

This is only copied code because it was not possible to call this method
because the method is private. This method should actually be public in CAS.
There should simply be no changes here.
This is not a great solution, but there is probably no better one at
the moment.