#99 Make validated dogu keys optional
The two keys are restricted to a set of values so that other values are not
allowed. When configuring these keys with `cesapp edit-config cas` the UX
is terribly bad because the user is forced to manually enter an already set
value.

This commit marks these keys as `optional` even though they are mandatory.
Anyhow the `edit-config` behaviour holds still the mandatoryness true by
validating the keys and thus forcing a value from the given set.