Add the attribute for the ID of the user #133
This value is read out in the LdapAuthenticationProperties
class in the field 'principalAttributeId'.

The value must be set so that CAS uses the ID from the LDAP
entry for the user. If the value is not set, this leads to
problems, as in this case the username entered by the user
is used. This can differ in spelling (upper and lower case).