Configure activation of password reset via e-mail #156
You can now configure whether the functionality for a password
reset via a link sent by e-mail should be activated.

Note: In order to be able to recognise in the Thymeleaf
template, there is a new entry in the custom-messages
properties file that stores whether the password reset feature
is active or not. This is a simple workaround.