Suppress determination of an existing username #161
Suppress determination of an existing username via password reset function.

Up to now, an error message has appeared if a user name does not exist in
the system. If the user name is available in the system, a confirmation
follows that an e-mail has been sent.

Now a success message is displayed even if a user does not exist in the
system.