Add integration tests #161
In order to better reach certain elements, "data-testid"s have been added
to the HTML files.

In the test step 'CAS shows a confirmation page about a sent email',
unfortunately concrete texts have to be checked because the elements on
the page do not have unique IDs. Adding "data-testid "s is not possible
here, because the pages have not been adapted for CES (and an adaptation
only for the integration tests does not make sense).