#173 avoid in-container FS log files, write access logs into Stdout instead
This commit removes a bug which writes log files to the in-container filesystem.
The AccessLog component in question uses its own logging system and
configuration. Thus, changing the config key `logging/root` or its mapping does
not change the behaviour, so it must be configured differently.

While it is possible to include other access log components that directly print
out the access logs to Stdout, this commit's solution does not include any
further libraries.

Instead, the access log component still writes to a file but this time to
`/dev/stdout`. Both, the suffix must be set to the empty string and the
rotation must be turned off in order to avoid a generated file name other than
`/dev/stdout`. A suffix would lead to `/dev/stdout.${suffix}`, a rotation might
lead to `/dev/stdout${_DD_MM_and_so_on}`.