#198 fully rename throttling keys, add proper doc sentences
The previous key naming is no longer useful because the semantics of the
single throttling keys have changed. This commit renames the key
limit/max_number to limit/failure_threshold which is vital for enabling
throttling in the first place.

The key limit/failure_store_time renames to limit/stale_removal_interval
to reflect that only throttled accounts are removed from the in-memory
map that went stale according the limit/lock_time (aka
throttle-window-seconds).

Co-authored-by: Nico Franzeck <nico.franzeck@cloudogu.com>