#246 accept empty config value for `oidc/enabled`
The previous templating relied on this key to be explicitly set, either
to false or to true. This is not always the case if the key is untouched
by the user. This should map to false.

This commit inverts the logic as trade-in for a slightly more
light-weight statement compared to
`{{ if eq (.Config.GetOrDefault "false" "oidc/enabled") "true"}}...`