#297 Revert LDAP dependency to be an optional dependency again.
User directories like LDAP and Active Directory are first class citizens
in the CES set-up since a long time. Thus, the configuration values are
positioned in a way that they _not really_ allow for a double
configuration: external directory AND internal LDAP dogu.

Instead, the values of the external directory must match to provide
steady access to the directory. The LDAP dogu might then catch up on
config values from the external directory, leading to dogu start up
errors.

Besides being a technical problem this problem also arises another, both
aesthetic and architectural: Administrators should not be forced to
install two different solutions to the same problem to solve a problem
that only arises from a problem within a unsufficient MN component
implementation.

This commit reverts the ldap dogu dependency to an optional one and
updates the necessary documents. This will provide admins with a way to
an successful CAS upgrade again.